+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ### Fixed
 
 - `machine` directive: support non-numeric _cpu_type_ (`mc68010`)
+- `reg` directive: support register mask constant
 
 ## 0.1.9 - 2025-09-08